Extract `AppLocaleProviderGateway` to into a separate file
