+++ lib/sharezone_localizations/lib/sharezone_localizations.dart
@@ -1,4 +1,5 @@
 export 'localizations/sharezone_localizations.gen.dart';
+export 'src/app_locale_gateway.dart';
 export 'src/app_locale_provider.dart';
 export 'src/app_locales.dart';
 export 'src/context_extension.dart';
